Several changes made:
Minor modifications to style.

publish & created_at attributes added per item. Items with the publish
attribute set to false will not be processed & output.

Made a start on a couple of posts, though they are both WIPs.

A change to the layout means that the navbar now has automated output,
along with them now being sorted.

Minor changes to styles, including the addition of a new font for code
elements.