Several changes.
Creation of a weather project page.
Creation of a post detailing R function call semantics.
The weather data empty db is updated to use the new schema, it is also
being unpublished to ensure that it is not going to overwrite what is
running live at the moment.
Small modification to the GH contest entry post.
Updated the index page to not be so cheesy.
Updated the layout to include links to friends.