A large amount of content was modified due to the update to nanoc 3.1.x.
Due to the update, binary assets have been implemented. This means no
odd file extensions for binary data, nor any associated metadata.

When it comes to sorting content now, the sort_by method is used. This
is simply because it's more concise and gets the point across better
than the sort method that was used previously. The <=> operator is no
longer needed to define the sort method.

The Rules file was improved to use less hard coding for the routing.